requirements/base.txt
bcrypt
databases[postgresql]
fastapi
sqlalchemy
python-jose[cryptography]
python-multipart
pydantic[email]